Add test to favorite number exercise
Before adding the new test it was possible to pass all tests by just
returning the value of favorite_list[:number] if it's not nil and
42 if it's nil. This solution leads to the unwanted behaviour that it
returns 42 when the :number key is present but has nil as value.
The new test tests this edge case so that the solution above does not
pass all tests anymore.